+++ controller/main.php
@@ -1,7 +1,7 @@
 <?php
 /**
 *
-* @package Usermap v0.8.x
+* @package Usermap v0.9.x
 * @copyright (c) 2020 Mike-on-Tour
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
@@ -78,7 +78,8 @@ class main
 		/*
 		*	get configuration values to send them to javascript for initialising the map
 		*/
-		$usermap_config = $this->config['mot_usermap_lat']."|".$this->config['mot_usermap_lon']."|".$this->config['mot_usermap_zoom'];
+		$usermap_config = $this->config['mot_usermap_lat'] . '|' .$this->config['mot_usermap_lon'] . '|' . $this->config['mot_usermap_zoom'] . '|' .
+						$this->config['mot_usermap_markers_pc'] . '|' . $this->config['mot_usermap_markers_mob'];
 		$server_config = $this->config['server_protocol'] . $this->config['server_name'] . $this->config['script_path'];
 		$poi_enabled = $this->config['mot_usermap_poi_enable'];
 
@@ -132,7 +133,7 @@ class main
 			$colour_text = ($row['group_colour']) ? ' style="color:#' . $row['group_colour'] . '"' : '';
 			$group_name = ($row['group_type'] == GROUP_SPECIAL) ? $this->language->lang('G_' . $row['group_name']) : $row['group_name'];
 
-			if ($row['group_name'] == 'BOTS' || ($this->user->data['user_id'] != ANONYMOUS && !$this->auth->acl_get('u_viewprofile')))
+			if ($this->user->data['user_id'] == ANONYMOUS && !$this->auth->acl_get('u_viewprofile'))
 			{
 				$usergroup_legend[] = '<span' . $colour_text . '>' . $group_name . '</span>';
 			}