+++ ext.php
@@ -2,7 +2,7 @@
 
 /**
 *
-* @package Usermap v0.8.0
+* @package Usermap v0.9.0
 * @copyright (c) 2019, 2020 Mike-on-Tour
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
@@ -15,6 +15,6 @@ class ext extends \phpbb\extension\base
 	public function is_enableable()
 	{
 		$config = $this->container->get('config');
-		return phpbb_version_compare($config['version'], '3.2.6', '>=');
+		return phpbb_version_compare($config['version'], '3.2.0', '>=');
 	}
 }